requirements.txt
tensorflow_gpu==2.3.0
matplotlib==3.4.1
scikit_image==0.17.2
numpy==1.18.1
easydict==1.9
pandas==1.1.4
Keras==2.4.3
tqdm==4.50.0
Pillow==8.2.0
scikit_learn==0.24.2
skimage==0.0
tensorflow==2.5.0
